# schemas/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: We Sell Houses API Documentation
  description: API for property management and communication between user and property agent
  contact:
    name: Filip Prosovsky
    email: prosovskyf@uni.coventry.ac.uk
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'https://observe-prime.codio-box.uk/api/v1'
    description: Development server for this application api
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer
...
paths:
...
    /properties:
      summary: Property view
      description: This resource manages view of properties
      parameters:
      - in: query
        name: direction
        schema:
          type: string
        required: false
        description: Order property by date created ASC/DESC
      - in: query
        name: features
        schema:
          type: string
        required: false
        description: Filter properties by features
      - in: query
        name: category
        schema:
          type: string
        required: false
        description: Filter properties by category
      get:
        tags:
          - Properties
        description:
          Get all public properties
        parameters:
          - in: query
            name: query
            schema:
              type: string
            required: false
            description: Query to filter properties
        responses:
          '200':
            description: Object with properties in array
            content:
              application/json:
                schema:
                  $ref: ./properties.json#/definitions/propertyView
          '404':
            description: Not found
    /properties/{id}:
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric id of the property
      get:
        tags:
          - Properties
        description:
          Get specific property by id
        responses:
          '200':
            description: Object with property information
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    property:
                      $ref: ./properties.json#/definitions/propertyView
                    agent:
                      $ref: ./agents.json#/definitions/agentProfile
          '400':
            description: Wrong id or format of id
          '404':
            description: Not found
    /admin/properties:
      summary: Admin properties
      description: This resource manages agent's property view and operations
      get:
        tags:
          - Properties
        description:
          Get all private properties for specific agent
        security:
          - BearerAuth: []
        parameters:
          - in: query
            name: direction
            schema:
              type: string
            required: false
            description: Order property by date created ASC/DESC
          - in: query
            name: features
            schema:
              type: string
            required: false
            description: Filter properties by features
          - in: query
            name: category
            schema:
              type: string
            required: false
            description: Filter properties by category
        responses:
          '200':
            description: Object with properties in array
            content:
              application/json:
                schema:
                  $ref: ./properties.json#/definitions/propertyView
          '403':
            description: Not authorized
          '404':
            description: Not found
      post:
        tags:
          - Properties
        description:
          Create property listing 
        security:
          - BearerAuth: []
        requestBody:
          description: Property information
          required: true
          content:
            application/json:
              schema:
                $ref: ./properties.json#/definitions/property
        responses:
          '201':
            description: Object with created property data
            content:
              application/json:
                schema:
                  $ref: ./properties.json#/definitions/propertyView
          '400':
            description: Wrong values entered
          '403':
            description: Not authorized
    /admin/properties/{id}:
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric id of the property
      get:
        tags:
          - Properties
        description:
          Get specific agent's property by id
        security:
          - BearerAuth: []
        responses:
          '200':
            description: Object with property information
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    property:
                      $ref: ./properties.json#/definitions/propertyView
          '403':
            description: Not authorized
          '404':
            description: Not found
      put:
        tags:
          - Properties
        description:
          Update specific property
        security:
          - BearerAuth: []
        requestBody:
          description: Property information
          required: true
          content:
            application/json:
              schema:
                $ref: ./properties.json#/definitions/property
        responses:
          '201':
            description: Property updated
          '400':
            description: Wrong values entered / Wrong ID
          '403':
            description: Not authorized
      delete:
        tags:
          - Properties
        description:
          Delete specific property, all images and features associated
        security:
          - BearerAuth: []
        responses:
          '201':
            description: Property deleted
          '400':
            description: Wrong ID input
          '403':
            description: Not authorized
    /admin/properties/hot:
      get:
        tags:
          - Properties
        description:
          Get properties marked as high priority
        security:
          - BearerAuth: []
        responses:
          '200':
            description: Object with property information
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    property:
                      $ref: ./properties.json#/definitions/propertyView
          '403':
            description: Not authorized
          '404':
            description: Not found
...
    /verification:
      summary: Verification
      description: >
        This resource manages account verification, password reset
        and change password chain after getting reset link
      get:
        tags:
          - Verification
        description:
          Verify activation token sent during signup
        parameters:
        - in: query
          name: username
          schema:
            type: string
            description: Username
          required: true
        - in: query
          name: token
          schema:
            type: string
            description: Verification token 
          required: true
        responses:
          '200':
            description: Account activated
            content:
              application/json:
                schema:
                  type: string
          "400":
            description: Not valid token
          '498':
            description: Token expired
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    text:
                      type: string
                    link: 
                      type: string
                      description: Generate link to resend new activation link to email
    /verification/resend:
      get:
        tags:
          - Verification
        description:
          Resend activation link to email
        parameters:
        - in: query
          name: username
          schema:
            type: string
            description: Username
          required: true
        responses:
          '200':
            description: Activation link sent, need to be confirmed in 48 hours
    /verification/resetpass:
      get:
        tags:
          - Verification
        description:
          Validate token after user clicked on reset link
        parameters:
        - in: query
          name: username
          schema:
            type: string
            description: Username
          required: true
        - in: query
          name: token
          schema:
            type: string
            description: Verification token
          required: true 
        responses:
          '200':
            description: Validated, fill password and will move to /verification/changepass resource
          "400":
            description: Not valid token
    /verification/changepass:
      post:
        tags:
          - Verification
        description:
          Change user entered password after reset
        parameters:
        - in: query
          name: username
          schema:
            type: string
            description: Username
          required: true
        - in: query
          name: token
          schema:
            type: string
            description: Verification token
          required: true 
        requestBody:
          description: Password
          required: true
          content:
            application/json:
              schema:
                properties:
                  secret:
                    type: string
                    description: Actual password
                required:
                  - secret
        responses:
          '200':
            description: Password changed, redirection to login
          '400':
            description: Not valid request
...